Move generateMockPncQueryResultFromAho to tests/helpers/comparison folder
