+++ packages/core/comparison/lib/comparePhase1.ts
@@ -13,8 +13,8 @@ import { extractExceptionsFromXml, parseAhoXml } from "../../lib/parse/parseAhoX
 import serialiseToXml from "../../lib/serialise/ahoXml/serialiseToXml"
 import getMessageType from "../../phase1/lib/getMessageType"
 import phase1Handler from "../../phase1/phase1"
+import generateMockPncQueryResultFromAho from "../../tests/helpers/comparison/generateMockPncQueryResultFromAho"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
-import generateMockPncQueryResultFromAho from "./generateMockPncQueryResultFromAho"
 import getPncQueryTimeFromAho from "./getPncQueryTimeFromAho"
 import isIntentionalDifference from "./isIntentionalDifference"
 import parseIncomingMessage from "./parseIncomingMessage"
