+++ packages/core/comparison/lib/comparePncMatching.ts
@@ -8,8 +8,8 @@ import type PncComparisonResultDetail from "../types/PncComparisonResultDetail"
 import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
 import { parseAhoXml } from "../../lib/parse/parseAhoXml"
 import CorePhase1 from "../../phase1/phase1"
+import generateMockPncQueryResultFromAho from "../../tests/helpers/comparison/generateMockPncQueryResultFromAho"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
-import generateMockPncQueryResultFromAho from "./generateMockPncQueryResultFromAho"
 import getPncQueryTimeFromAho from "./getPncQueryTimeFromAho"
 import parseIncomingMessage from "./parseIncomingMessage"
 import summariseMatching from "./summariseMatching"
