+++ packages/core/phase1/phase1.comparison.test.ts
@@ -10,7 +10,6 @@ import type { OldPhase1Comparison } from "../tests/types/ComparisonFile"
 import type Exception from "../types/Exception"
 import type Phase1Result from "./types/Phase1Result"
 
-import generateMockPncQueryResultFromAho from "../comparison/lib/generateMockPncQueryResultFromAho"
 import getPncQueryTimeFromAho from "../comparison/lib/getPncQueryTimeFromAho"
 import parseIncomingMessage from "../comparison/lib/parseIncomingMessage"
 import { sortExceptions } from "../comparison/lib/sortExceptions"
@@ -21,6 +20,7 @@ import { extractExceptionsFromXml } from "../lib/parse/parseAhoXml"
 import serialiseToXml from "../lib/serialise/ahoXml/serialiseToXml"
 import checkDatabaseMatches from "../tests/helpers/comparison/checkDatabaseMatches"
 import { clearDatabase, disconnectDb, sortTriggers, sql } from "../tests/helpers/comparison/e2eComparisonTestsHelpers"
+import generateMockPncQueryResultFromAho from "../tests/helpers/comparison/generateMockPncQueryResultFromAho"
 import getComparisonTests from "../tests/helpers/comparison/getComparisonTests"
 import MockPncGateway from "../tests/helpers/MockPncGateway"
 import phase1 from "./phase1"
